Revamped Controls/Game Manager
Controls now has a wrapper Key class that is used to house a particular
key and its delegates. Upon adding a fourth key, the realization came
that there would be 16 functions/delegates and more checks to be made
for one control script, so to make things easier the Key class will
house all of it. Controls looks so much cleaner now!

It now accounts for all key presses (tapping a key, holding a key, and
releasing a key). The only downside is that every minigame must add 12
more abstract functions, but this isn't really avoidable. However, you
do not have to put any logic for any of a key press, so just use the
functions you need and leave the rest alone.

Game manager now houses a dictionary of players, making it easier to
access the game object of the player and their stats class. These two
components are very important to making our game, but referencing it
didn't make much sense and calling it from game manager is just too
cluttered. This will hopefully make things cleaner.

You should now be able to use 4 keys as well. Please see the my sample
mini game for the relevant changes.