Added Fading Effects/Progress
Major Changes
-Fading effect added. A UI Manager is made to handle fades and other
recurring general UI for the game.
-Fades now work for starting a new game, and going to the progress
screen. For now, it is YOUR responsibility to call the fade functions
when transitioning from the instructions to the actual game.
-Instruction screen is done. Now, all minigames will not start unless
the player clicks Start on the instruction screen. Instruction screens
vary, but to keep things consistent copy the template that I provided in
this push.
-GameManager,Minigame, and the sample minigames have been modified to
reflect these changes.
NOTES:
-You MUST handle what happens when the player starts the game. That is,
initialize the game on Start(), but actually start the game when the
Start button is clicked (Minigame class may be changed to reflect these
requirements)
-You will need to use the UI Canvas in the Main Menu Scene in order to
get the fade effects to work. Otherwise, the game may break.
-Of course, you don't need the fade effects to test your game for now.
These are simply transitions that will be added in the end.