Added Animations/Updated Tags
-Added Jackie's animations onto the game. The TestMinigame utilizes the
animations, so you can see that for a sample.
-NOTE: To add in the animations the players need the model as a child of
them for it to work properly. There is a female mage prefab already
made; All you need to do is drag it into the player in your hierarchy
and call the animator functions (if you know how the animations work).
-Tags updated for John's game.
-Minigames now require a OnStart,OnRules, and OnControls function to
incorporate actually starting the game and removing the instructions
screen. Protected panel gameobjects are made too for all minigames to
utilize them for the instructions screen template.